+++ run-migrations.sql
@@ -1,5 +1,8 @@
 \set ON_ERROR_STOP on
 
+\echo Waiting for the FlowPilot schema migration lock
+SELECT pg_advisory_lock(hashtextextended('flowpilot-schema-migrations', 0));
+
 \echo Applying FlowPilot migration 001
 \ir postgres/init/001_create_leads.sql
 \echo Applying FlowPilot migration 002
@@ -16,4 +19,5 @@
 \ir postgres/init/007_add_partial_work_reconciliation.sql
 \echo Applying FlowPilot migration 008
 \ir postgres/init/008_add_reliability_observability.sql
+SELECT pg_advisory_unlock(hashtextextended('flowpilot-schema-migrations', 0));
 \echo FlowPilot migrations complete